+++ src/apps/agave/gnosis/agave.variable-borrow.token-fetcher.ts
@@ -33,6 +33,6 @@ export class GnosisAgaveVariableBorrowTokenFetcher extends AaveV2LendingTemplate
   }
 
   async getTertiaryLabel({ appToken }: GetDisplayPropsParams<AaveV2AToken, AaveV2LendingTokenDataProps>) {
-    return `${(appToken.dataProps.apy * 100).toFixed(3)}% (variable)`;
+    return `${appToken.dataProps.apy.toFixed(3)}% (variable)`;
   }
 }