+++ src/apps/geist/fantom/geist.stable-debt.token-fetcher.ts
@@ -33,6 +33,6 @@ export class FantomGeistStableDebtTokenFetcher extends AaveV2LendingTemplateToke
   }
 
   async getTertiaryLabel({ appToken }: GetDisplayPropsParams<AaveV2AToken, AaveV2LendingTokenDataProps>) {
-    return `${(appToken.dataProps.apy * 100).toFixed(3)}% APR (stable)`;
+    return `${appToken.dataProps.apy.toFixed(3)}% APR (stable)`;
   }
 }