+++ src/apps/geist/fantom/geist.supply.token-fetcher.ts
@@ -33,6 +33,6 @@ export class FantomGeistSupplyTokenFetcher extends AaveV2LendingTemplateTokenFet
   }
 
   async getTertiaryLabel({ appToken }: GetDisplayPropsParams<AaveV2AToken, AaveV2LendingTokenDataProps>) {
-    return `${(appToken.dataProps.apy * 100).toFixed(3)}% APY`;
+    return `${appToken.dataProps.apy.toFixed(3)}% APY`;
   }
 }