+++ src/apps/hector-network/contracts/ethers/index.ts
@@ -2,16 +2,14 @@
 /* tslint:disable */
 /* eslint-disable */
 export type { HectorNetworkBondDepository } from './HectorNetworkBondDepository';
-export type { HectorNetworkBscBondDepository } from './HectorNetworkBscBondDepository';
-export type { HectorNetworkFtmBondDepository } from './HectorNetworkFtmBondDepository';
+export type { HectorNetworkBondNoTreasury } from './HectorNetworkBondNoTreasury';
 export type { HectorNetworkStakeBondDepository } from './HectorNetworkStakeBondDepository';
 export type { HectorNetworkStaked } from './HectorNetworkStaked';
 export type { HectorNetworkStakingRewards } from './HectorNetworkStakingRewards';
 export type { HectorNetworkToken } from './HectorNetworkToken';
 export * as factories from './factories';
 export { HectorNetworkBondDepository__factory } from './factories/HectorNetworkBondDepository__factory';
-export { HectorNetworkBscBondDepository__factory } from './factories/HectorNetworkBscBondDepository__factory';
-export { HectorNetworkFtmBondDepository__factory } from './factories/HectorNetworkFtmBondDepository__factory';
+export { HectorNetworkBondNoTreasury__factory } from './factories/HectorNetworkBondNoTreasury__factory';
 export { HectorNetworkStakeBondDepository__factory } from './factories/HectorNetworkStakeBondDepository__factory';
 export { HectorNetworkStaked__factory } from './factories/HectorNetworkStaked__factory';
 export { HectorNetworkStakingRewards__factory } from './factories/HectorNetworkStakingRewards__factory';